requirements.txt
flask
pyopenssl
cryptography
pyjwt
Gunicorn
